refactor: callManager
