+++ logic/src/androidMain/kotlin/com/wire/kalium/logic/feature/call/scenario/ParticipantChangedHandlerImpl.kt
@@ -1,4 +1,4 @@
-package com.wire.kalium.logic.feature.call.callback
+package com.wire.kalium.logic.feature.call.scenario
 
 import com.sun.jna.Pointer
 import com.wire.kalium.calling.callbacks.ParticipantChangedHandler
