+++ logic/src/jvmMain/kotlin/com/wire/kalium/logic/feature/call/CallManagerImpl.kt
@@ -9,12 +9,7 @@ import com.wire.kalium.logic.kaliumLogger
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.asStateFlow
 
-val calls = MutableStateFlow(listOf<Call>())
-
 actual class CallManagerImpl : CallManager {
-
-    private val _calls = MutableStateFlow(listOf<Call>())
-    override val allCalls = _calls.asStateFlow()
 
     init {
         kaliumLogger.w("CallManager initialized for JVM but not supported yet.")
